refactor(stream): Move to stream export of Needed
