+++ src/stream/bstr.rs
@@ -1,11 +1,11 @@
 use core::num::NonZeroUsize;
 
-use crate::error::Needed;
 use crate::stream::AsBStr;
 use crate::stream::Checkpoint;
 use crate::stream::Compare;
 use crate::stream::CompareResult;
 use crate::stream::FindSlice;
+use crate::stream::Needed;
 use crate::stream::Offset;
 #[cfg(feature = "unstable-recover")]
 #[cfg(feature = "std")]
